% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{fun.GetCombinationFromList}
\alias{fun.GetCombinationFromList}
\title{Load in unexport function for Np dynamic reconstruction}
\usage{
fun.GetCombinationFromList(ListIn)
}
\description{
Load in unexport function for Np dynamic reconstruction
}
