% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNpBootrap.R
\name{GetNpDynamic_90Sampling}
\alias{GetNpDynamic_90Sampling}
\title{Bootrapping with 90\% Sampling}
\usage{
GetNpDynamic_90Sampling(FFFly, TTTreeAnn, TTTipInfo, FFFilterOrgan)
}
\description{
Bootrapping with 90\% Sampling
}
