% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDownSamplingNp.R
\name{GetNpDynamic_DownSampling}
\alias{GetNpDynamic_DownSampling}
\title{Assess sampling sufficiency for each organ}
\usage{
GetNpDynamic_DownSampling(FFFly, TTTreeAnn, TTTipInfo, FFFilterOrgan)
}
\description{
Assess sampling sufficiency for each organ
}
