% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCoalescentSim_VarMutRate.R
\name{GetNpDynamic_Sim_VarMutRate}
\alias{GetNpDynamic_Sim_VarMutRate}
\title{Simulation of coalescent method under different mutation rates}
\usage{
GetNpDynamic_Sim_VarMutRate(DDData, MMMutRate)
}
\description{
Simulation of coalescent method under different mutation rates
}
