% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNpVarMutRate.R
\name{GetNpDynamic_VarMutRate}
\alias{GetNpDynamic_VarMutRate}
\title{Reconstruct under variant mutation rates}
\usage{
GetNpDynamic_VarMutRate(FFFly, TTTreeAnn, TTTipInfo, FFFilterOrgan)
}
\description{
Reconstruct under variant mutation rates
}
