% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub__GetTreeWithNpDynamic.R
\name{GetNpDynamicTable}
\alias{GetNpDynamicTable}
\title{Subfunctions for Build Tree with Np dynamic}
\usage{
GetNpDynamicTable(DDDynamic)
}
\description{
Subfunctions for Build Tree with Np dynamic
}
