% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCoalescentSim_PopStructure.R
\name{GetSimTree_Panmictic}
\alias{GetSimTree_Panmictic}
\title{Simulation of coalescent method with different population structure}
\usage{
GetSimTree_Panmictic(FFFile, DDDynamic_1, DDDynamic_2, MMMutRate)
}
\description{
Simulation of coalescent method with different population structure
}
