% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub__RunCoalescentSim_VarMutRate.R
\name{GetSubTreeWithNode2Height}
\alias{GetSubTreeWithNode2Height}
\title{Subfunctions for RunCoalescentSim_VarMutRate.R}
\usage{
GetSubTreeWithNode2Height(TTTree, NNNode2Height, RRRatio)
}
\description{
Subfunctions for RunCoalescentSim_VarMutRate.R
}
