% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPackgedData.R
\name{GetTreeAnn}
\alias{GetTreeAnn}
\title{extract Np dynamic from *.newick and save as PackagedData.RData}
\usage{
GetTreeAnn(FFFly)
}
\description{
extract Np dynamic from *.newick and save as PackagedData.RData
}
